Changing the location of Discount folder and remove unused imports
