+++ src/main/java/com/SEGroup/Domain/Discount/Conditions/AndCondition.java
@@ -1,23 +1,23 @@
-package com.SEGroup.Domain.Store.Discount.Conditions;
+package com.SEGroup.Domain.Discount.Conditions;
 
 import com.SEGroup.Domain.ProductCatalog.StoreSearchEntry;
 
 import java.util.List;
 import java.util.function.Predicate;
 
-public class OrCondition extends CompositeCondition{
+public class AndCondition extends CompositeCondition {
 
-    public OrCondition(List<Predicate<StoreSearchEntry[]>> conditions){
+    public AndCondition(List<Predicate<StoreSearchEntry[]>> conditions){
         super(conditions);
     }
 
     @Override
     public boolean test(StoreSearchEntry[] entries) {
         for (Predicate<StoreSearchEntry[]> condition : conditions) {
-            if (condition.test(entries)) {
-                return true;
+            if (!condition.test(entries)) {
+                return false;
             }
         }
-        return false;
+        return true;
     }
 }
