+++ src/main/java/com/SEGroup/Domain/Discount/Numerical/NumericalComposite.java
@@ -1,7 +1,7 @@
-package com.SEGroup.Domain.Store.Discount.Numerical;
+package com.SEGroup.Domain.Discount.Numerical;
 
 import com.SEGroup.Domain.ProductCatalog.StoreSearchEntry;
-import com.SEGroup.Domain.Store.Discount.Discount;
+import com.SEGroup.Domain.Discount.Discount;
 import com.SEGroup.Infrastructure.Repositories.InMemoryProductCatalog;
 
 import java.util.List;
