+++ src/main/java/com/SEGroup/Domain/Discount/SimpleDiscount.java
@@ -1,4 +1,4 @@
-package com.SEGroup.Domain.Store.Discount;
+package com.SEGroup.Domain.Discount;
 
 import com.SEGroup.Domain.ProductCatalog.StoreSearchEntry;
 import com.SEGroup.Infrastructure.Repositories.InMemoryProductCatalog;
