+++ src/test/java/com/SEGroup/UnitTests/Discount/Conditions/AndConditionTest.java
@@ -1,13 +1,13 @@
-package com.SEGroup.UnitTests.StoreTests.Discount.Conditions;
+package com.SEGroup.UnitTests.Discount.Conditions;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import com.SEGroup.Domain.ProductCatalog.StoreSearchEntry;
-import com.SEGroup.Domain.Store.Discount.ConditionalDiscount;
-import com.SEGroup.Domain.Store.Discount.Conditions.AndCondition;
-import com.SEGroup.Domain.Store.Discount.Discount;
-import com.SEGroup.Domain.Store.Discount.DiscountScope;
-import com.SEGroup.Domain.Store.Discount.SimpleDiscount;
+import com.SEGroup.Domain.Discount.ConditionalDiscount;
+import com.SEGroup.Domain.Discount.Conditions.AndCondition;
+import com.SEGroup.Domain.Discount.Discount;
+import com.SEGroup.Domain.Discount.DiscountScope;
+import com.SEGroup.Domain.Discount.SimpleDiscount;
 import com.SEGroup.Infrastructure.Repositories.InMemoryProductCatalog;
 import org.junit.jupiter.api.Test;
