+++ src/test/java/com/SEGroup/UnitTests/Discount/DiscountScopeTest.java
@@ -1,7 +1,7 @@
-package com.SEGroup.UnitTests.StoreTests.Discount;
+package com.SEGroup.UnitTests.Discount;
 
 import com.SEGroup.Domain.ProductCatalog.StoreSearchEntry;
-import com.SEGroup.Domain.Store.Discount.DiscountScope;
+import com.SEGroup.Domain.Discount.DiscountScope;
 import com.SEGroup.Infrastructure.Repositories.InMemoryProductCatalog;
 import org.junit.jupiter.api.Test;
